+++ ci.py
@@ -79,12 +79,12 @@ class GitlabCi(GitlabCiBuilderImpl):
             build_pipeline_job = (
                 self.job(BuildTriggerPipeline)
                 .apply()
-                .with_name(f"{test_name}/build-pipeline")
                 .with_tags(["static-k8s"])
+                .add_to_name(f" -f pipelines/{test_name}/ci.py")
             )
-            build_pipeline_job.model.script = [
-                f'exec sh -c "gcip2 build-pipeline -f pipelines/{test_name}/ci.py"',
-            ]
+            # build_pipeline_job.model.script = [
+            #     f'exec sh -c "gcip2 build-pipeline -f pipelines/{test_name}/ci.py"',
+            # ]
 
             trigger_pipeline_job = (
                 self.job(TriggerPipeline)