+++ gcip2/tasks_core/ci_tasks/__init__.py
@@ -1,0 +1,3 @@
+from gcip2.tasks_core.ci_tasks._tasks import CI_TASK_GENERATOR
+
+__all__ = ("CI_TASK_GENERATOR",)