+++ gcip2/tasks_core/ci_tasks/_params.py
@@ -1,0 +1,21 @@
+from gcip2.tasks_core.models.params import Param
+
+
+def ci_file() -> Param:
+    return Param(
+        name="ci_file",
+        long="ci-file",
+        short="f",
+        default="ci.py",
+        type=str,
+    )
+
+
+def out_gitlab_ci(default: str) -> Param:
+    return Param(
+        name="out_pipeline",
+        long="out-pipeline",
+        short="o",
+        default=default,
+        type=str,
+    )